refactor: split self_improvement.py into focused modules
Extract 3 sections (646 lines) from self_improvement.py (2122→1496 lines):
- rule_pipeline.py (259 lines): PipelineResult, skill generator, run_rule_pipeline
- causal_chains.py (117 lines): CausalRelation, CausalLink, CausalChain
- clustering.py (140 lines): RuleCluster, detect_contradictions, cluster_rules

All imports updated in session_close.py, _core.py, and test files.
Backward-compat __getattr__ in self_improvement.py for any missed imports.
2831 tests passing, zero behavior change.
