+++ src/gradata/enhancements/rule_pipeline.py
@@ -1,0 +1,382 @@
+"""Rule Pipeline Orchestrator — 3-phase rule pipeline adapted from Hindsight retain.
+
+SDK LAYER: Layer 1 (enhancements). Imports from self_improvement and shared types.
+
+Extracted from self_improvement.py to keep that module focused on the core
+graduation logic (confidence scoring, FSRS, graduation state machine).
+"""
+
+from __future__ import annotations
+
+import logging
+import os
+import re
+from dataclasses import dataclass, field
+from pathlib import Path
+
+from gradata._types import LessonState
+
+_log = logging.getLogger(__name__)
+
+
+@dataclass
+class PipelineResult:
+    """Result of a full rule pipeline run."""
+
+    graduated: list[str] = field(default_factory=list)
+    demoted: list[str] = field(default_factory=list)
+    meta_rules_created: list[str] = field(default_factory=list)
+    hooks_promoted: list[str] = field(default_factory=list)
+    disposition_updates: dict[str, dict] = field(default_factory=dict)
+    freshness_updates: int = 0
+    errors: list[str] = field(default_factory=list)
+    skills_generated: list[str] = field(default_factory=list)
+
+
+def _generate_skill_file(
+    lesson: "object",
+    output_dir: "Path",
+) -> "Path | None":
+    """Generate a SKILL.md file from a graduated rule.
+
+    Only generates for rules meeting quality gate:
+    - State == RULE
+    - Confidence >= 0.90
+    - fire_count >= 3
+
+    Args:
+        lesson: The lesson to convert into a skill file.
+        output_dir: Root directory under which per-skill subdirs are created.
+
+    Returns:
+        Path to the written SKILL.md, or None if the lesson doesn't qualify.
+    """
+    if not hasattr(lesson, "state") or lesson.state.name != "RULE":
+        return None
+    if lesson.confidence < 0.90 or getattr(lesson, "fire_count", 0) < 3:
+        return None
+
+    # Survival events = correction_event_ids that survived (proxy: fire_count)
+    survival_count = len(getattr(lesson, "correction_event_ids", []) or [])
+
+    slug = re.sub(
+        r"[^a-z0-9]+",
+        "-",
+        f"{lesson.category}-{lesson.description[:40]}".lower(),
+    ).strip("-")
+    skill_dir = output_dir / slug
+    skill_dir.mkdir(parents=True, exist_ok=True)
+
+    skill_path = skill_dir / "SKILL.md"
+
+    content = f"""---
+name: {lesson.description[:60]}
+description: Auto-graduated from correction-driven learning (confidence {lesson.confidence:.2f}, fired {getattr(lesson, 'fire_count', 0)} times)
+source: gradata-behavioral-engine
+confidence: {lesson.confidence}
+category: {lesson.category}
+graduated_at_session: {getattr(lesson, 'created_session', 0)}
+---
+
+# {lesson.description}
+
+**Category**: {lesson.category}
+**Confidence**: {lesson.confidence:.2f}
+**Times Applied**: {getattr(lesson, 'fire_count', 0)}
+
+## Directive
+
+{lesson.description}
+
+## Context
+
+This rule was learned from {survival_count} user corrections and graduated through INSTINCT -> PATTERN -> RULE stages with ablation validation.
+
+## When to Apply
+
+Apply this directive when working on tasks related to: {lesson.category}
+"""
+
+    skill_path.write_text(content, encoding="utf-8")
+    return skill_path
+
+
+def review_generated_skill(skill_path: "Path") -> dict:
+    """Review a generated skill file for quality issues.
+
+    Returns dict with:
+    - valid: bool
+    - issues: list[str] - problems found
+    - suggestions: list[str] - improvements
+    - path: str - absolute path of the reviewed file
+    """
+    import re as _re
+
+    text = skill_path.read_text(encoding="utf-8")
+    issues: list[str] = []
+    suggestions: list[str] = []
+
+    # Check frontmatter exists
+    if not text.startswith("---"):
+        issues.append("Missing frontmatter")
+
+    # Check minimum content length
+    if len(text) < 100:
+        issues.append("Skill too short (< 100 chars)")
+
+    # Check for placeholder text
+    if "(requires" in text.lower() or "todo" in text.lower():
+        issues.append("Contains placeholder text")
+
+    # Check confidence is reasonable
+    conf_match = _re.search(r"confidence:\s*([\d.]+)", text)
+    if conf_match:
+        conf = float(conf_match.group(1))
+        if conf < 0.90:
+            issues.append(f"Low confidence ({conf:.2f}) - below RULE threshold")
+
+    # Check description isn't too vague
+    desc_match = _re.search(r"description:\s*(.+)", text)
+    if desc_match:
+        desc = desc_match.group(1).strip()
+        if len(desc) < 20:
+            suggestions.append("Description is very short - consider expanding")
+        if desc.startswith("Auto-graduated"):
+            suggestions.append("Description is generic - consider adding specific context")
+
+    # Check directive section exists and has content
+    if "## Directive" not in text:
+        issues.append("Missing Directive section")
+    elif text.split("## Directive")[1].split("##")[0].strip() == "":
+        issues.append("Directive section is empty")
+
+    return {
+        "valid": len(issues) == 0,
+        "issues": issues,
+        "suggestions": suggestions,
+        "path": str(skill_path),
+    }
+
+
+def run_rule_pipeline(
+    lessons_path: Path,
+    db_path: Path,
+    current_session: int,
+    corrections: list[dict] | None = None,
+) -> PipelineResult:
+    """Coordinated 3-phase rule pipeline.
+
+    Adapted from Hindsight's 3-phase retain orchestrator:
+    - Phase 1 (read): Load state, compute freshness, rank rules
+    - Phase 2 (atomic): Graduate rules, create meta-rules, update confidence
+    - Phase 3 (best-effort): Hook promotion, disposition updates, events
+
+    Args:
+        lessons_path: Path to lessons.md file.
+        db_path: Path to system.db.
+        current_session: Current session number.
+        corrections: Recent corrections from this session.
+
+    Returns:
+        PipelineResult with all changes made.
+    """
+    from gradata.enhancements.self_improvement import (
+        MIN_APPLICATIONS_FOR_PATTERN,
+        MIN_APPLICATIONS_FOR_RULE,
+        PATTERN_THRESHOLD,
+        RULE_THRESHOLD,
+        format_lessons,
+        parse_lessons,
+    )
+
+    result = PipelineResult()
+    corrections = corrections or []
+
+    # ── Phase 1: Read (non-destructive) ──────────────────────────────────────
+    # Load all state needed for decision-making. No writes.
+    try:
+        text = lessons_path.read_text(encoding="utf-8")
+        all_lessons = parse_lessons(text)
+    except Exception as exc:
+        result.errors.append(f"Phase 1: failed to load lessons: {exc}")
+        return result
+
+    # Compute freshness for all graduated rules
+    try:
+        from gradata.enhancements.freshness import (  # type: ignore[import]
+            Trend,
+            compute_trend,
+        )
+
+        for lesson in all_lessons:
+            if lesson.state.name in ("RULE", "PATTERN"):
+                correction_sessions = [
+                    {"session": current_session - i, "severity": "minor"}
+                    for i in range(getattr(lesson, "fire_count", 0))
+                ]
+                trend = compute_trend(correction_sessions, current_session)
+                if trend == Trend.STALE:
+                    sessions_stale = getattr(lesson, "sessions_since_fire", 0)
+                    if sessions_stale > 30:
+                        decay = -0.01 * (sessions_stale - 29)
+                        lesson.confidence = max(0.0, lesson.confidence + decay)
+                        result.freshness_updates += 1
+    except ImportError:
+        pass  # freshness module not available
+
+    # Rank rules using retrieval fusion if available
+    try:
+        from gradata.enhancements.retrieval_fusion import (  # type: ignore[import]
+            ScoredRule,
+            apply_correction_boost,
+            reciprocal_rank_fusion,
+        )
+
+        scored_by_confidence = [
+            ScoredRule(
+                rule_id=getattr(l, "description", "")[:40],
+                text=l.description,
+                score=l.confidence,
+                source="confidence",
+                metadata={
+                    "from_correction": bool(getattr(l, "correction_event_ids", None)),
+                    "recency_score": 0.8 if l.state.name == "RULE" else 0.5,
+                    "severity_score": 0.5,
+                },
+            )
+            for l in all_lessons
+            if l.state.name in ("RULE", "PATTERN") and l.confidence >= 0.60
+        ]
+        if scored_by_confidence:
+            merged = reciprocal_rank_fusion([scored_by_confidence])
+            apply_correction_boost(merged)
+            _log.debug("Phase 1: ranked %d rules via fusion", len(merged))
+    except ImportError:
+        pass  # retrieval_fusion not available
+
+    # ── Phase 2: Atomic writes ────────────────────────────────────────────────
+    # Graduate rules, update confidence, create meta-rules.
+    for lesson in all_lessons:
+        if lesson.state.name == "INSTINCT" and lesson.confidence >= PATTERN_THRESHOLD:
+            if lesson.fire_count >= MIN_APPLICATIONS_FOR_PATTERN:
+                lesson.state = LessonState.PATTERN
+                result.graduated.append(f"{lesson.category}:{lesson.description[:30]}")
+        elif lesson.state.name == "PATTERN" and lesson.confidence >= RULE_THRESHOLD:
+            if lesson.fire_count >= MIN_APPLICATIONS_FOR_RULE:
+                lesson.state = LessonState.RULE
+                result.graduated.append(f"{lesson.category}:{lesson.description[:30]}")
+
+    # Synthesize meta-rules from graduated rules
+    try:
+        from gradata.enhancements.meta_rules import synthesize_meta_rules_agentic  # type: ignore[import]
+        from gradata.enhancements.meta_rules_storage import (  # type: ignore[import]
+            load_meta_rules,
+            save_meta_rules,
+        )
+
+        existing_metas = []
+        if db_path.is_file():
+            existing_metas = load_meta_rules(db_path)
+
+        new_metas = synthesize_meta_rules_agentic(
+            lessons=all_lessons,
+            existing_metas=existing_metas,
+            current_session=current_session,
+        )
+        if new_metas and db_path.is_file():
+            save_meta_rules(db_path, existing_metas + new_metas)
+            result.meta_rules_created = [m.id for m in new_metas]
+    except (ImportError, Exception) as exc:
+        result.errors.append(f"Phase 2: meta-rule synthesis: {exc}")
+
+    # Write lessons back
+    try:
+        lessons_path.write_text(format_lessons(all_lessons), encoding="utf-8")
+    except Exception as exc:
+        result.errors.append(f"Phase 2: failed to write lessons: {exc}")
+
+    # ── Phase 3: Best-effort (non-critical) ───────────────────────────────────
+    # Hook promotion, disposition updates, event emission.
+    # Failures here are logged but don't abort the pipeline.
+
+    # Hook promotion for newly graduated RULE-state lessons
+    try:
+        from gradata.enhancements.rule_to_hook import classify_rule, promote  # type: ignore[import]
+
+        for lesson in all_lessons:
+            if lesson.state.name == "RULE" and lesson.confidence >= RULE_THRESHOLD:
+                candidate = classify_rule(lesson.description, lesson.confidence)
+                if candidate.determinism.value != "not_deterministic":
+                    try:
+                        gen_result = promote(
+                            lesson.description,
+                            lesson.confidence,
+                            lesson=lesson,
+                            source="pipeline",
+                        )
+                        if getattr(gen_result, "installed", False):
+                            result.hooks_promoted.append(lesson.description[:40])
+                    except Exception as exc:
+                        result.errors.append(f"Phase 3: hook promotion: {exc}")
+    except ImportError:
+        pass
+
+    # Disposition updates from this session's corrections
+    try:
+        from gradata.enhancements.behavioral_engine import DispositionTracker  # type: ignore[import]
+
+        tracker = DispositionTracker()
+        disp_path = lessons_path.parent / "disposition.json"
+        if disp_path.is_file():
+            import json as _json
+            tracker = DispositionTracker.from_dict(
+                _json.loads(disp_path.read_text(encoding="utf-8"))
+            )
+        for correction in corrections:
+            category = correction.get("category", "")
+            severity = correction.get("severity", "minor")
+            domain = correction.get("domain", "global")
+            disp = tracker.update_from_correction(domain, category, severity)
+            if domain not in result.disposition_updates:
+                result.disposition_updates[domain] = {
+                    "skepticism": disp.skepticism,
+                    "literalism": disp.literalism,
+                    "empathy": disp.empathy,
+                }
+        if result.disposition_updates:
+            try:
+                import json as _json
+                disp_path.write_text(
+                    _json.dumps(tracker.to_dict(), indent=2), encoding="utf-8",
+                )
+            except Exception as exc:
+                result.errors.append(f"Phase 3: disposition write: {exc}")
+    except ImportError:
+        pass
+
+    # Skill generation for high-confidence rules (best-effort, env-gated)
+    if os.environ.get("GRADATA_ENABLE_SKILL_EXPORT"):
+        try:
+            skills_dir = lessons_path.parent.parent / ".claude" / "skills" / "generated"
+            for lesson in all_lessons:
+                skill_path = _generate_skill_file(lesson, skills_dir)
+                if skill_path:
+                    review = review_generated_skill(skill_path)
+                    if not review["valid"]:
+                        skill_path.unlink(missing_ok=True)
+                        result.errors.append(f"Phase 3: skill rejected: {review['issues']}")
+                    else:
+                        result.skills_generated.append(str(skill_path))
+        except Exception as exc:
+            result.errors.append(f"Phase 3: skill generation: {exc}")
+
+    _log.info(
+        "Pipeline complete: %d graduated, %d meta-rules, %d hooks, %d freshness updates, %d skills, %d errors",
+        len(result.graduated),
+        len(result.meta_rules_created),
+        len(result.hooks_promoted),
+        result.freshness_updates,
+        len(result.skills_generated),
+        len(result.errors),
+    )
+    return result
