+++ src/gradata/enhancements/self_improvement.py
@@ -1475,648 +1475,22 @@ def compute_learning_velocity(
 
 
 # ---------------------------------------------------------------------------
-# Rule Pipeline Orchestrator (3-phase, adapted from Hindsight retain)
+# Backward-compat re-exports (symbols moved to focused modules)
 # ---------------------------------------------------------------------------
 
 
-@dataclass
-class PipelineResult:
-    """Result of a full rule pipeline run."""
-
-    graduated: list[str] = field(default_factory=list)
-    demoted: list[str] = field(default_factory=list)
-    meta_rules_created: list[str] = field(default_factory=list)
-    hooks_promoted: list[str] = field(default_factory=list)
-    disposition_updates: dict[str, dict] = field(default_factory=dict)
-    freshness_updates: int = 0
-    errors: list[str] = field(default_factory=list)
-    skills_generated: list[str] = field(default_factory=list)
-
-
-def _generate_skill_file(
-    lesson: "Lesson",
-    output_dir: "Path",
-) -> "Path | None":
-    """Generate a SKILL.md file from a graduated rule.
-
-    Only generates for rules meeting quality gate:
-    - State == RULE
-    - Confidence >= 0.90
-    - fire_count >= 3
-
-    Args:
-        lesson: The lesson to convert into a skill file.
-        output_dir: Root directory under which per-skill subdirs are created.
-
-    Returns:
-        Path to the written SKILL.md, or None if the lesson doesn't qualify.
-    """
-    if not hasattr(lesson, "state") or lesson.state.name != "RULE":
-        return None
-    if lesson.confidence < 0.90 or getattr(lesson, "fire_count", 0) < 3:
-        return None
-
-    # Survival events = correction_event_ids that survived (proxy: fire_count)
-    survival_count = len(getattr(lesson, "correction_event_ids", []) or [])
-
-    slug = re.sub(
-        r"[^a-z0-9]+",
-        "-",
-        f"{lesson.category}-{lesson.description[:40]}".lower(),
-    ).strip("-")
-    skill_dir = output_dir / slug
-    skill_dir.mkdir(parents=True, exist_ok=True)
-
-    skill_path = skill_dir / "SKILL.md"
-
-    content = f"""---
-name: {lesson.description[:60]}
-description: Auto-graduated from correction-driven learning (confidence {lesson.confidence:.2f}, fired {getattr(lesson, 'fire_count', 0)} times)
-source: gradata-behavioral-engine
-confidence: {lesson.confidence}
-category: {lesson.category}
-graduated_at_session: {getattr(lesson, 'created_session', 0)}
----
-
-# {lesson.description}
-
-**Category**: {lesson.category}
-**Confidence**: {lesson.confidence:.2f}
-**Times Applied**: {getattr(lesson, 'fire_count', 0)}
-
-## Directive
-
-{lesson.description}
-
-## Context
-
-This rule was learned from {survival_count} user corrections and graduated through INSTINCT -> PATTERN -> RULE stages with ablation validation.
-
-## When to Apply
-
-Apply this directive when working on tasks related to: {lesson.category}
-"""
-
-    skill_path.write_text(content, encoding="utf-8")
-    return skill_path
-
-
-def review_generated_skill(skill_path: "Path") -> dict:
-    """Review a generated skill file for quality issues.
-
-    Returns dict with:
-    - valid: bool
-    - issues: list[str] - problems found
-    - suggestions: list[str] - improvements
-    - path: str - absolute path of the reviewed file
-    """
-    text = skill_path.read_text(encoding="utf-8")
-    issues: list[str] = []
-    suggestions: list[str] = []
-
-    # Check frontmatter exists
-    if not text.startswith("---"):
-        issues.append("Missing frontmatter")
-
-    # Check minimum content length
-    if len(text) < 100:
-        issues.append("Skill too short (< 100 chars)")
-
-    # Check for placeholder text
-    if "(requires" in text.lower() or "todo" in text.lower():
-        issues.append("Contains placeholder text")
-
-    # Check confidence is reasonable
-    conf_match = re.search(r"confidence:\s*([\d.]+)", text)
-    if conf_match:
-        conf = float(conf_match.group(1))
-        if conf < 0.90:
-            issues.append(f"Low confidence ({conf:.2f}) - below RULE threshold")
-
-    # Check description isn't too vague
-    desc_match = re.search(r"description:\s*(.+)", text)
-    if desc_match:
-        desc = desc_match.group(1).strip()
-        if len(desc) < 20:
-            suggestions.append("Description is very short - consider expanding")
-        if desc.startswith("Auto-graduated"):
-            suggestions.append("Description is generic - consider adding specific context")
-
-    # Check directive section exists and has content
-    if "## Directive" not in text:
-        issues.append("Missing Directive section")
-    elif text.split("## Directive")[1].split("##")[0].strip() == "":
-        issues.append("Directive section is empty")
-
-    return {
-        "valid": len(issues) == 0,
-        "issues": issues,
-        "suggestions": suggestions,
-        "path": str(skill_path),
-    }
-
-
-def run_rule_pipeline(
-    lessons_path: Path,
-    db_path: Path,
-    current_session: int,
-    corrections: list[dict] | None = None,
-) -> PipelineResult:
-    """Coordinated 3-phase rule pipeline.
-
-    Adapted from Hindsight's 3-phase retain orchestrator:
-    - Phase 1 (read): Load state, compute freshness, rank rules
-    - Phase 2 (atomic): Graduate rules, create meta-rules, update confidence
-    - Phase 3 (best-effort): Hook promotion, disposition updates, events
-
-    Args:
-        lessons_path: Path to lessons.md file.
-        db_path: Path to system.db.
-        current_session: Current session number.
-        corrections: Recent corrections from this session.
-
-    Returns:
-        PipelineResult with all changes made.
-    """
-    result = PipelineResult()
-    corrections = corrections or []
-
-    # ── Phase 1: Read (non-destructive) ──────────────────────────────────────
-    # Load all state needed for decision-making. No writes.
-    try:
-        text = lessons_path.read_text(encoding="utf-8")
-        all_lessons = parse_lessons(text)
-    except Exception as exc:
-        result.errors.append(f"Phase 1: failed to load lessons: {exc}")
-        return result
-
-    # Compute freshness for all graduated rules
-    try:
-        from gradata.enhancements.freshness import (  # type: ignore[import]
-            Trend,
-            compute_trend,
-        )
-
-        for lesson in all_lessons:
-            if lesson.state.name in ("RULE", "PATTERN"):
-                correction_sessions = [
-                    {"session": current_session - i, "severity": "minor"}
-                    for i in range(getattr(lesson, "fire_count", 0))
-                ]
-                trend = compute_trend(correction_sessions, current_session)
-                if trend == Trend.STALE:
-                    sessions_stale = getattr(lesson, "sessions_since_fire", 0)
-                    if sessions_stale > 30:
-                        decay = -0.01 * (sessions_stale - 29)
-                        lesson.confidence = max(0.0, lesson.confidence + decay)
-                        result.freshness_updates += 1
-    except ImportError:
-        pass  # freshness module not available
-
-    # Rank rules using retrieval fusion if available
-    try:
-        from gradata.enhancements.retrieval_fusion import (  # type: ignore[import]
-            ScoredRule,
-            apply_correction_boost,
-            reciprocal_rank_fusion,
-        )
-
-        scored_by_confidence = [
-            ScoredRule(
-                rule_id=getattr(l, "description", "")[:40],
-                text=l.description,
-                score=l.confidence,
-                source="confidence",
-                metadata={
-                    "from_correction": bool(getattr(l, "correction_event_ids", None)),
-                    "recency_score": 0.8 if l.state.name == "RULE" else 0.5,
-                    "severity_score": 0.5,
-                },
-            )
-            for l in all_lessons
-            if l.state.name in ("RULE", "PATTERN") and l.confidence >= 0.60
-        ]
-        if scored_by_confidence:
-            merged = reciprocal_rank_fusion([scored_by_confidence])
-            apply_correction_boost(merged)
-            _log.debug("Phase 1: ranked %d rules via fusion", len(merged))
-    except ImportError:
-        pass  # retrieval_fusion not available
-
-    # ── Phase 2: Atomic writes ────────────────────────────────────────────────
-    # Graduate rules, update confidence, create meta-rules.
-    for lesson in all_lessons:
-        if lesson.state.name == "INSTINCT" and lesson.confidence >= PATTERN_THRESHOLD:
-            if lesson.fire_count >= MIN_APPLICATIONS_FOR_PATTERN:
-                lesson.state = LessonState.PATTERN
-                result.graduated.append(f"{lesson.category}:{lesson.description[:30]}")
-        elif lesson.state.name == "PATTERN" and lesson.confidence >= RULE_THRESHOLD:
-            if lesson.fire_count >= MIN_APPLICATIONS_FOR_RULE:
-                lesson.state = LessonState.RULE
-                result.graduated.append(f"{lesson.category}:{lesson.description[:30]}")
-
-    # Synthesize meta-rules from graduated rules
-    try:
-        from gradata.enhancements.meta_rules import synthesize_meta_rules_agentic  # type: ignore[import]
-        from gradata.enhancements.meta_rules_storage import (  # type: ignore[import]
-            load_meta_rules,
-            save_meta_rules,
-        )
-
-        existing_metas = []
-        if db_path.is_file():
-            existing_metas = load_meta_rules(db_path)
-
-        new_metas = synthesize_meta_rules_agentic(
-            lessons=all_lessons,
-            existing_metas=existing_metas,
-            current_session=current_session,
-        )
-        if new_metas and db_path.is_file():
-            save_meta_rules(db_path, existing_metas + new_metas)
-            result.meta_rules_created = [m.id for m in new_metas]
-    except (ImportError, Exception) as exc:
-        result.errors.append(f"Phase 2: meta-rule synthesis: {exc}")
-
-    # Write lessons back
-    try:
-        lessons_path.write_text(format_lessons(all_lessons), encoding="utf-8")
-    except Exception as exc:
-        result.errors.append(f"Phase 2: failed to write lessons: {exc}")
-
-    # ── Phase 3: Best-effort (non-critical) ───────────────────────────────────
-    # Hook promotion, disposition updates, event emission.
-    # Failures here are logged but don't abort the pipeline.
-
-    # Hook promotion for newly graduated RULE-state lessons
-    try:
-        from gradata.enhancements.rule_to_hook import classify_rule, promote  # type: ignore[import]
-
-        for lesson in all_lessons:
-            if lesson.state.name == "RULE" and lesson.confidence >= RULE_THRESHOLD:
-                candidate = classify_rule(lesson.description, lesson.confidence)
-                if candidate.determinism.value != "not_deterministic":
-                    try:
-                        gen_result = promote(
-                            lesson.description,
-                            lesson.confidence,
-                            lesson=lesson,
-                            source="pipeline",
-                        )
-                        if getattr(gen_result, "installed", False):
-                            result.hooks_promoted.append(lesson.description[:40])
-                    except Exception as exc:
-                        result.errors.append(f"Phase 3: hook promotion: {exc}")
-    except ImportError:
-        pass
-
-    # Disposition updates from this session's corrections
-    try:
-        from gradata.enhancements.behavioral_engine import DispositionTracker  # type: ignore[import]
-
-        tracker = DispositionTracker()
-        disp_path = lessons_path.parent / "disposition.json"
-        if disp_path.is_file():
-            import json as _json
-            tracker = DispositionTracker.from_dict(
-                _json.loads(disp_path.read_text(encoding="utf-8"))
-            )
-        for correction in corrections:
-            category = correction.get("category", "")
-            severity = correction.get("severity", "minor")
-            domain = correction.get("domain", "global")
-            disp = tracker.update_from_correction(domain, category, severity)
-            if domain not in result.disposition_updates:
-                result.disposition_updates[domain] = {
-                    "skepticism": disp.skepticism,
-                    "literalism": disp.literalism,
-                    "empathy": disp.empathy,
-                }
-        if result.disposition_updates:
-            try:
-                import json as _json
-                disp_path.write_text(
-                    _json.dumps(tracker.to_dict(), indent=2), encoding="utf-8",
-                )
-            except Exception as exc:
-                result.errors.append(f"Phase 3: disposition write: {exc}")
-    except ImportError:
-        pass
-
-    # Skill generation for high-confidence rules (best-effort, env-gated)
-    import os
-    if os.environ.get("GRADATA_ENABLE_SKILL_EXPORT"):
-        try:
-            skills_dir = lessons_path.parent.parent / ".claude" / "skills" / "generated"
-            for lesson in all_lessons:
-                skill_path = _generate_skill_file(lesson, skills_dir)
-                if skill_path:
-                    review = review_generated_skill(skill_path)
-                    if not review["valid"]:
-                        skill_path.unlink(missing_ok=True)
-                        result.errors.append(f"Phase 3: skill rejected: {review['issues']}")
-                    else:
-                        result.skills_generated.append(str(skill_path))
-        except Exception as exc:
-            result.errors.append(f"Phase 3: skill generation: {exc}")
-
-    _log.info(
-        "Pipeline complete: %d graduated, %d meta-rules, %d hooks, %d freshness updates, %d skills, %d errors",
-        len(result.graduated),
-        len(result.meta_rules_created),
-        len(result.hooks_promoted),
-        result.freshness_updates,
-        len(result.skills_generated),
-        len(result.errors),
-    )
-    return result
-
-
-# ---------------------------------------------------------------------------
-# Causal Chain Tracking (adapted from Hindsight causal relations)
-# ---------------------------------------------------------------------------
-
-
-class CausalRelation(StrEnum):
-    CORRECTION_TO_RULE = "correction_to_rule"  # Correction produced this rule
-    RULE_TO_BEHAVIOR = "rule_to_behavior"  # Rule changed this behavior
-    CORRECTION_CHAIN = "correction_chain"  # Corrections share root cause
-    REINFORCEMENT = "reinforcement"  # Correction reinforced existing rule
-
-
-@dataclass
-class CausalLink:
-    """A causal relationship between two events in the learning pipeline."""
-
-    source_id: str  # Event/correction/rule that caused this
-    target_id: str  # Event/correction/rule that was caused
-    relation: CausalRelation  # Type of causal link
-    strength: float = 1.0  # 0.0-1.0 confidence in the link
-    session: int = 0
-
-
-class CausalChain:
-    """Tracks causal links between corrections, rules, and behaviors."""
-
-    def __init__(self) -> None:
-        self._links: list[CausalLink] = []
-
-    def add_link(
-        self,
-        source_id: str,
-        target_id: str,
-        relation: CausalRelation,
-        strength: float = 1.0,
-        session: int = 0,
-    ) -> CausalLink:
-        link = CausalLink(source_id, target_id, relation, strength, session)
-        self._links.append(link)
-        return link
-
-    def trace_rule_origin(self, rule_id: str) -> list[CausalLink]:
-        """Trace backward: what corrections produced this rule?"""
-        return [
-            l
-            for l in self._links
-            if l.target_id == rule_id
-            and l.relation in (CausalRelation.CORRECTION_TO_RULE, CausalRelation.REINFORCEMENT)
-        ]
-
-    def trace_rule_impact(self, rule_id: str) -> list[CausalLink]:
-        """Trace forward: what behaviors did this rule change?"""
-        return [
-            l
-            for l in self._links
-            if l.source_id == rule_id and l.relation == CausalRelation.RULE_TO_BEHAVIOR
-        ]
-
-    def find_correction_chains(self, correction_id: str) -> list[CausalLink]:
-        """Find corrections sharing a root cause with this one."""
-        return [
-            l
-            for l in self._links
-            if (l.source_id == correction_id or l.target_id == correction_id)
-            and l.relation == CausalRelation.CORRECTION_CHAIN
-        ]
-
-    def get_rule_provenance(self, rule_id: str) -> dict:
-        """Full provenance for a rule: corrections that built it + behaviors it changed.
-
-        This is the data structure ablation testing needs to prove causation.
-        """
-        origins = self.trace_rule_origin(rule_id)
-        impacts = self.trace_rule_impact(rule_id)
-        return {
-            "rule_id": rule_id,
-            "correction_sources": [
-                {"id": l.source_id, "strength": l.strength, "session": l.session}
-                for l in origins
-            ],
-            "behavioral_impacts": [
-                {"id": l.target_id, "strength": l.strength, "session": l.session}
-                for l in impacts
-            ],
-            "total_evidence": len(origins) + len(impacts),
-        }
-
-    def to_list(self) -> list[dict]:
-        """Serialize all links."""
-        return [
-            {
-                "source_id": l.source_id,
-                "target_id": l.target_id,
-                "relation": l.relation.value,
-                "strength": l.strength,
-                "session": l.session,
-            }
-            for l in self._links
-        ]
-
-    @classmethod
-    def from_list(cls, data: list[dict]) -> CausalChain:
-        """Deserialize from list of dicts."""
-        chain = cls()
-        for item in data:
-            chain._links.append(
-                CausalLink(
-                    source_id=item["source_id"],
-                    target_id=item["target_id"],
-                    relation=CausalRelation(item["relation"]),
-                    strength=item.get("strength", 1.0),
-                    session=item.get("session", 0),
-                )
-            )
-        return chain
-
-    @property
-    def link_count(self) -> int:
-        return len(self._links)
-
-
-# ---------------------------------------------------------------------------
-# Phase 1: Rule Clustering & Contradiction Detection
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class RuleCluster:
-    """A cluster of related rules with shared domain/category."""
-
-    cluster_id: str
-    domain: str
-    category: str
-    member_ids: list[str] = field(default_factory=list)
-    cluster_confidence: float = 0.0
-    summary: str = ""
-    contradictions: list[tuple[str, str]] = field(default_factory=list)
-
-    @property
-    def size(self) -> int:
-        return len(self.member_ids)
-
-    @property
-    def has_contradictions(self) -> bool:
-        return len(self.contradictions) > 0
-
-
-def detect_contradictions(
-    lessons: list[Lesson],
-) -> list[tuple[str, str]]:
-    """Find potentially contradicting rules within a group.
-
-    Uses keyword overlap + negation detection. Two rules contradict if:
-    - They share significant keyword overlap (>= 3 common tokens)
-    - One contains negation language the other doesn't
-
-    Returns list of (rule_id_1, rule_id_2) pairs.
-    """
-    _NEGATION_WORDS = {"never", "dont", "don", "not", "no", "stop", "avoid", "without"}
-
-    def tokenize(text: str) -> set[str]:
-        return set(re.findall(r"[a-z]{3,}", text.lower()))
-
-    def has_negation(tokens: set[str]) -> bool:
-        return bool(tokens & _NEGATION_WORDS)
-
-    contradictions: list[tuple[str, str]] = []
-    for i, a in enumerate(lessons):
-        tokens_a = tokenize(a.description)
-        neg_a = has_negation(tokens_a)
-
-        for b in lessons[i + 1 :]:
-            tokens_b = tokenize(b.description)
-            neg_b = has_negation(tokens_b)
-
-            # Need significant overlap AND different negation status
-            overlap = len(tokens_a & tokens_b)
-            if overlap >= 3 and neg_a != neg_b:
-                id_a = f"{a.category}:{a.description[:40]}"
-                id_b = f"{b.category}:{b.description[:40]}"
-                contradictions.append((id_a, id_b))
-
-    return contradictions
-
-
-def cluster_rules(
-    lessons: list[Lesson],
-    min_cluster_size: int = 2,
-) -> list[RuleCluster]:
-    """Group graduated rules into clusters by category.
-
-    Rules in the same category with similar descriptions are clustered.
-    Cluster confidence = weighted mean of member confidences.
-    """
-    import json
-    from collections import defaultdict
-
-    # Only cluster RULE and PATTERN tier
-    graduated = [l for l in lessons if l.state.name in ("RULE", "PATTERN")]
-
-    # Group by category
-    by_category: dict[str, list[Lesson]] = defaultdict(list)
-    for lesson in graduated:
-        by_category[lesson.category].append(lesson)
-
-    clusters: list[RuleCluster] = []
-    for category, members in by_category.items():
-        if len(members) < min_cluster_size:
-            continue
-
-        member_ids = [f"{m.category}:{m.description[:40]}" for m in members]
-        avg_conf = sum(m.confidence for m in members) / len(members)
-
-        # Detect contradictions within cluster
-        contradictions = detect_contradictions(members)
-
-        # Build summary from member descriptions
-        descriptions = [m.description for m in members[:5]]
-        summary = f"{len(members)} rules in {category}: " + "; ".join(descriptions)
-        if len(members) > 5:
-            summary += f" (+{len(members) - 5} more)"
-
-        domain = ""
-        if members[0].scope_json:
-            try:
-                scope = json.loads(members[0].scope_json)
-                domain = scope.get("domain", "")
-            except (json.JSONDecodeError, TypeError):
-                pass
-
-        cluster = RuleCluster(
-            cluster_id=f"cluster-{category.lower()}",
-            domain=domain or "global",
-            category=category,
-            member_ids=member_ids,
-            cluster_confidence=round(avg_conf, 4),
-            summary=summary,
-            contradictions=contradictions,
-        )
-        clusters.append(cluster)
-
-    return clusters
-
-
-def promote_instinct_clusters(
-    lessons: list[Lesson],
-    min_cluster_size: int = 3,
-    coherence_threshold: float = 0.80,
-) -> list[str]:
-    """Find INSTINCT-tier clusters that deserve PATTERN promotion.
-
-    3+ INSTINCT rules in the same category with high inter-coherence
-    (no contradictions, similar confidence) get promoted as a group.
-
-    Returns list of promoted lesson descriptions.
-    """
-    from collections import defaultdict
-
-    instinct = [l for l in lessons if l.state.name == "INSTINCT"]
-    by_category: dict[str, list[Lesson]] = defaultdict(list)
-    for lesson in instinct:
-        by_category[lesson.category].append(lesson)
-
-    promoted: list[str] = []
-    for category, members in by_category.items():
-        if len(members) < min_cluster_size:
-            continue
-
-        contradictions = detect_contradictions(members)
-        if contradictions:
-            continue  # Incoherent cluster, skip
-
-        # Check confidence coherence (std dev < threshold complement)
-        confs = [m.confidence for m in members]
-        mean_conf = sum(confs) / len(confs)
-        variance = sum((c - mean_conf) ** 2 for c in confs) / len(confs)
-        std_dev = variance ** 0.5
-
-        if std_dev > (1.0 - coherence_threshold):
-            continue  # Too much variance
-
-        # Promote all members to PATTERN
-        for member in members:
-            member.state = LessonState.PATTERN
-            member.confidence = max(member.confidence, 0.60)
-            promoted.append(member.description)
-
-    return promoted
+def __getattr__(name: str):  # type: ignore[return]
+    _PIPELINE_NAMES = {"PipelineResult", "run_rule_pipeline", "_generate_skill_file", "review_generated_skill"}
+    _CAUSAL_NAMES = {"CausalRelation", "CausalLink", "CausalChain"}
+    _CLUSTER_NAMES = {"RuleCluster", "detect_contradictions", "cluster_rules", "promote_instinct_clusters"}
+
+    if name in _PIPELINE_NAMES:
+        from gradata.enhancements import rule_pipeline
+        return getattr(rule_pipeline, name)
+    if name in _CAUSAL_NAMES:
+        from gradata.enhancements import causal_chains
+        return getattr(causal_chains, name)
+    if name in _CLUSTER_NAMES:
+        from gradata.enhancements import clustering
+        return getattr(clustering, name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
